move lab post form to components
